Added custom hook for handling http errors
