+++ src/hoc/withErrorHandler/withErrorHandler.js
@@ -1,36 +1,14 @@
-import React, { useState, useEffect } from "react";
+import React from "react";
 import Modal from "../../components/UI/Modal/Modal";
+import useHttpErrorHandler from '../../hooks/http-error-handler';
 import Aux from "../AuxHoc";
 
 const withErrorHandler = (WrappedComponent, axios) => {
     return (props) => {
-        const [error, setError] = useState(false);
-
-        const requestInterceptor = axios.interceptors.request.use((req) => {
-            setError(null);
-            return req;
-        });
-        const responseInterceptor = axios.interceptors.response.use(
-            (response) => response,
-            (error) => {
-                setError(error);
-                return Promise.reject(error);
-            }
-        );
-
-        useEffect(() => {
-            return () => {
-                axios.interceptors.request.eject(requestInterceptor);
-                axios.interceptors.response.eject(responseInterceptor);
-            };
-        }, [requestInterceptor, responseInterceptor]);
-
-        const errorConfirmedHandler = () => {
-            setError(null);
-        };
+        const [error, clearError] = useHttpErrorHandler(axios);
         return (
             <Aux>
-                <Modal show={error} modalClosed={errorConfirmedHandler}>
+                <Modal show={error} modalClosed={clearError}>
                     {error ? error.message : null}
                 </Modal>
                 <WrappedComponent {...props} />
